Split up consolidation module
